Fix OperandFactory

+++ OperandFactory.cpp
@@ -38,20 +38,20 @@ IOperand const *OperandFactory::createInt8(std::string const& value) const
 
 IOperand const *OperandFactory::createInt16(std::string const& value) const
 {
-	return new Operand<int16_t>(boost::numeric_cast<int8_t>(std::stoll(value)), eOperandType::Int8);
+	return new Operand<int16_t>(boost::numeric_cast<int8_t>(std::stoll(value)), eOperandType::Int16);
 }
 
 IOperand const *OperandFactory::createInt32(std::string const& value) const
 {
-	return new Operand<int32_t>(boost::numeric_cast<int32_t>(std::stoll(value)), eOperandType::Int8);
+	return new Operand<int32_t>(boost::numeric_cast<int32_t>(std::stoll(value)), eOperandType::Int32);
 }
 
 IOperand const *OperandFactory::createFloat(std::string const& value) const
 {
-	return new Operand<float>(std::stof(value), eOperandType::Int8);
+	return new Operand<float>(std::stof(value), eOperandType::Float);
 }
 
 IOperand const *OperandFactory::createDouble(std::string const& value) const
 {
-	return new Operand<double>(std::stod(value), eOperandType::Int8);
+	return new Operand<double>(std::stod(value), eOperandType::Double);
 }
